separate post-publish script
