+++ package.json
@@ -5,7 +5,8 @@
     "nodeclean": "rm -r node_modules && rm -r packages/*/node_modules",
     "preinstall": "npx only-allow pnpm",
     "changeset-version": "cd packages/ops && turbo build && pnpm run changeset-version",
-    "publish-all": "cd packages/ops && turbo build && pnpm run publish-all"
+    "publish-all": "cd packages/ops && turbo build && pnpm run publish-all",
+    "post-publish": "cd packages/ops && turbo build && pnpm run post-publish"
   },
   "devDependencies": {
     "@changesets/cli": "^2.26.2",
